avoid deletes on search_name in reverse-only mode

+++ sql/functions/placex_triggers.sql
@@ -715,7 +715,9 @@ BEGIN
                   and ( relation_members[i+1] != 'side_stream' or NEW.name->'name' = name->'name')
                 LOOP
                   UPDATE placex SET linked_place_id = NEW.place_id WHERE place_id = linked_node_id;
-                  DELETE FROM search_name WHERE place_id = linked_node_id;
+                  IF NOT %REVERSE-ONLY% THEN
+                    DELETE FROM search_name WHERE place_id = linked_node_id;
+		  END IF;
                 END LOOP;
               END IF;
           END LOOP;
@@ -858,7 +860,9 @@ BEGIN
     UPDATE placex set linked_place_id = NEW.place_id
       WHERE place_id = location.place_id;
     -- ensure that those places are not found anymore
-    DELETE FROM search_name WHERE place_id = location.place_id;
+    IF NOT %REVERSE-ONLY% THEN
+      DELETE FROM search_name WHERE place_id = location.place_id;
+    END IF;
 
     SELECT wikipedia, importance
       FROM compute_importance(location.extratags, NEW.country_code,
